turd: Use memcpy() to copy fixed-length unterminated strings
The "readonly" string pointer points to the start of the PHPTURD
environment variable value.  The "readonly" part of the string is
terminated by a colon rather than a NUL; this is why we have
"readonly_len" as a variable.

Use memcpy() to copy the "readonly" string to the destination buffer,
since there is effectively no NUL terminator.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>