turd: Avoid unnecessary string duplication
Change the internal API so that potentially turdified paths are to be
freed only if they differ (in pointer value) from the original path.
This allows the calls to strdup() to be elided completely, thereby
minimising the overhead of using the LD_PRELOAD library for paths
outside of the turd directories.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>