turd: Check for PHPTURD environment variable only once
There seems to be no way to have php-fpm vary the environment on a
per-request basis.  Given this, we may as well check the existence and
contents of PHPTURD only on the first invocation of a turdified
library call and then cache the result.

The copy of the PHPTURD environment variable is held permanently for
the lifetime of the process.  This will show up as a memory leak if
run under a tool such as Valgrind.  This is not a major concern, since
running any PHP code under Valgrind produces a truly mind-boggling
report listing hundreds of instances of leaked memory and undefined
behaviour in the PHP interpreter.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>